use int64 for `request_id`
mpv uses int64 for `request_id` internally[1]. Right now, once
`lastRequest` goes past 2**63 - 1, mpv will complain about `request_id`
not being an integer:

> 'request_id' must be an integer. Using other types is deprecated and
> will trigger an error in the future!

and will respond with `request_id` that looks like a float (e.g.
`9223372036854775808.000000`). We fail to parse it as `uint` and thus
match the response to the request, which causes the request to hang.

If we use `int64`, `lastRequest` will wrap after going past 2**63 - 1
and continue to increase from the negative side.

[1]: https://github.com/mpv-player/mpv/blob/2a969a833f98b3486be8007972887c27c08b6f61/input/ipc.c#L124

+++ mpvipc.go
@@ -16,8 +16,8 @@ type Connection struct {
 	client     net.Conn
 	socketName string
 
-	lastRequest     uint
-	waitingRequests map[uint]chan *commandResult
+	lastRequest     int64
+	waitingRequests map[int64]chan *commandResult
 
 	eventHub        *Hub
 
@@ -100,7 +100,7 @@ func NewConnection(socketName string) *Connection {
 	return &Connection{
 		socketName:      socketName,
 		lock:            &sync.Mutex{},
-		waitingRequests: make(map[uint]chan *commandResult),
+		waitingRequests: make(map[int64]chan *commandResult),
 		closeWaiters:    make(map[uint]chan struct{}),
 	}
 }
@@ -247,7 +247,7 @@ func (c *Connection) WaitUntilClosed() {
 	c.lock.Unlock()
 }
 
-func (c *Connection) sendCommand(id uint, arguments ...interface{}) error {
+func (c *Connection) sendCommand(id int64, arguments ...interface{}) error {
 	if c.client == nil {
 		return fmt.Errorf("trying to send command on closed mpv client")
 	}
@@ -272,13 +272,13 @@ func (c *Connection) sendCommand(id uint, arguments ...interface{}) error {
 
 type commandRequest struct {
 	Arguments []interface{} `json:"command"`
-	ID        uint          `json:"request_id"`
+	ID        int64         `json:"request_id"`
 }
 
 type commandResult struct {
 	Status string      `json:"error"`
 	Data   interface{} `json:"data"`
-	ID     uint        `json:"request_id"`
+	ID     int64       `json:"request_id"`
 }
 
 func (c *Connection) checkResult(data []byte) {
